+++ src/Plugin.php
@@ -28,8 +28,27 @@ class Plugin extends \craft\base\Plugin
         }
 
         if ( Craft::$app->request->isCpRequest && Craft::$app->getUser()->getIdentity() ) {
+            $view = Craft::$app->getView();
+            $scripts = $this->getSettings()->scripts;
+            $styles = $this->getSettings()->styles;
             // the includeJs method lets us add js to the bottom of the page
-            Craft::$app->getView()->registerJsVar('FAListIcons', $this->getSettings()->icons, View::POS_HEAD);
+            $view->registerJsVar('FAListIcons', $this->getSettings()->icons, View::POS_HEAD);
+
+            //adds Font Awesome kit script
+            if(!empty($scripts)){
+                foreach($scripts as $script){
+                    if(is_array($script) && isset($script['src'])){
+                        $view->registerJsFile($script['src'], isset($script['params']) ? $script['params'] : []);
+                    }
+                    elseif(is_string($script)) $view->registerJsFile($script);
+                }
+            }
+            //adds Font Awesome font styles
+            if(!empty($styles)){
+                foreach($styles as $style){
+                    $view->registerCssFile($style);
+                }
+            }
         }
     }